config: add default configure file
Currently the configuration items are to be defined by users, somehow
we should provide the configureation item list and explanations for
user to choose. So we add a default configuration file for user's
information, and user just have to define item that they want to
override in their own config file.

+++ sbin/functions.sh
@@ -301,7 +301,7 @@ validate_harddrive()
 
 	if [ -z ${harddrive_device} ] || [ -z ${harddrive_model} ]
 	then
-		debugmsg ${DEBUG_CRIT} "ERROR: Input parameters not provided"
+		debugmsg ${DEBUG_CRIT} "ERROR: validate_harddrive input parameters not provided"
 		return 1
 	fi
 
@@ -332,7 +332,7 @@ remove_partitions()
 
 	if [ -z $device ]
 	then
-		debugmsg ${DEBUG_CRIT} "ERROR: Input parameters not provided"
+		debugmsg ${DEBUG_CRIT} "ERROR: remove_partitions input parameters not provided"
 	fi
 
 	debugmsg ${DEBUG_INFO} "Removing all partitions"
@@ -369,7 +369,7 @@ create_partition()
 
 	if [ -z $device ]
 	then
-		debugmsg ${DEBUG_CRIT} "ERROR: Input parameters not provided"
+		debugmsg ${DEBUG_CRIT} "ERROR: create_partition input parameters not provided"
 	fi
 
 	debugmsg ${DEBUG_INFO} "Creating partition ${device}${partnum}"
@@ -403,7 +403,7 @@ create_filesystem()
 
 	if [ -z $partition ]
 	then
-		debugmsg ${DEBUG_CRIT} "ERROR: Input parameters not provided"
+		debugmsg ${DEBUG_CRIT} "ERROR: create_filesystem input parameters not provided"
 		return 1
 	fi
 
@@ -503,7 +503,7 @@ umount_partitions()
 
 	if [ -z $device ]
 	then
-		debugmsg ${DEBUG_CRIT} "ERROR: Input parameters not provided"
+		debugmsg ${DEBUG_CRIT} "ERROR: umount_partitions input parameters not provided"
 	fi
 
 	sync
@@ -622,7 +622,7 @@ install_grub()
 	
 	#install efi boot
 
-	if [ -n "${INSTALL_EFIBOOT}" ]; then
+	if [ -n "${INSTALL_EFIBOOT}" ] && [ -e "${INSTALL_EFIBOOT}" ]; then
 		debugmsg ${DEBUG_INFO} "Installing the EFI bootloader"
 		mkdir -p ${mountpoint}/EFI/BOOT/
 		cp $INSTALL_EFIBOOT ${mountpoint}/EFI/BOOT/
